# src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: Account Service API
  description: API para la gestión de cuentas bancarias
  version: 1.0.0
servers:
  - url: http://localhost:8082
    description: Servidor de desarrollo
  - url: http://api-gateway:8080/accounts
    description: API Gateway

tags:
  - name: accounts
    description: Operaciones con cuentas bancarias

paths:
  /accounts:
    get:
      tags:
        - accounts
      summary: Obtener todas las cuentas
      description: Retorna un listado de todas las cuentas registradas
      operationId: getAllAccounts
      responses:
        '200':
          description: Operación exitosa
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountDto'
    post:
      tags:
        - accounts
      summary: Crear una nueva cuenta
      description: Crea una nueva cuenta bancaria con los datos proporcionados
      operationId: createAccount
      requestBody:
        description: Datos de la cuenta a crear
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountDto'
        required: true
      responses:
        '201':
          description: Cuenta creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDto'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Cliente no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /accounts/{id}:
    get:
      tags:
        - accounts
      summary: Obtener cuenta por ID
      description: Retorna una cuenta específica según su ID
      operationId: getAccountById
      parameters:
        - name: id
          in: path
          description: ID de la cuenta
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Operación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDto'
        '404':
          description: Cuenta no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - accounts
      summary: Actualizar cuenta
      description: Actualiza los datos de una cuenta existente
      operationId: updateAccount
      parameters:
        - name: id
          in: path
          description: ID de la cuenta
          required: true
          schema:
            type: string
      requestBody:
        description: Datos de la cuenta a actualizar
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountDto'
        required: true
      responses:
        '200':
          description: Cuenta actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDto'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Cuenta no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - accounts
      summary: Eliminar cuenta
      description: Elimina una cuenta según su ID
      operationId: deleteAccount
      parameters:
        - name: id
          in: path
          description: ID de la cuenta
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Cuenta eliminada
        '404':
          description: Cuenta no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /accounts/customer/{customerId}:
    get:
      tags:
        - accounts
      summary: Obtener cuentas por cliente
      description: Retorna todas las cuentas asociadas a un cliente
      operationId: getAccountsByCustomerId
      parameters:
        - name: customerId
          in: path
          description: ID del cliente
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Operación exitosa
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountDto'

  /accounts/{id}/balance:
    get:
      tags:
        - accounts
      summary: Consultar saldo de cuenta
      description: Retorna el saldo actual de una cuenta
      operationId: getAccountBalance
      parameters:
        - name: id
          in: path
          description: ID de la cuenta
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Operación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceDto'
        '404':
          description: Cuenta no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /accounts/{id}/balance:
    put:
      tags:
        - accounts
      summary: Actualizar saldo de cuenta
      description: Actualiza el saldo de una cuenta (usado internamente por el servicio de transacciones)
      operationId: updateAccountBalance
      parameters:
        - name: id
          in: path
          description: ID de la cuenta
          required: true
          schema:
            type: string
        - name: amount
          in: query
          description: Monto a actualizar (positivo para depósitos, negativo para retiros)
          required: true
          schema:
            type: number
            format: double
      responses:
        '200':
          description: Saldo actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDto'
        '400':
          description: Error en la operación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Cuenta no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    AccountDto:
      type: object
      properties:
        id:
          type: string
          example: "60c73def2aa3a75f5c5d0f7c"
        accountNumber:
          type: string
          example: "ACC-12345678"
        type:
          type: string
          enum: [SAVINGS, CHECKING, FIXED_TERM]
          example: "SAVINGS"
        customerId:
          type: string
          example: "60c73def2aa3a75f5c5d0f7b"
        customerType:
          type: string
          enum: [PERSONAL, BUSINESS]
          example: "PERSONAL"
        balance:
          type: number
          format: double
          example: 1000.50
        maxMonthlyMovements:
          type: integer
          format: int32
          example: 20
        currentMonthlyMovements:
          type: integer
          format: int32
          example: 5
        withdrawalDay:
          type: string
          format: date
          example: "2023-07-15"
        holders:
          type: array
          items:
            type: string
          example: ["60c73def2aa3a75f5c5d0f7b"]
        signatories:
          type: array
          items:
            type: string
          example: ["60c73def2aa3a75f5c5d0f7d"]
        maintenanceFee:
          type: boolean
          example: false
      required:
        - type
        - customerId
        - customerType
        - balance

    BalanceDto:
      type: object
      properties:
        accountId:
          type: string
          example: "60c73def2aa3a75f5c5d0f7c"
        accountNumber:
          type: string
          example: "ACC-12345678"
        accountType:
          type: string
          example: "SAVINGS"
        balance:
          type: number
          format: double
          example: 1000.50
        remainingMonthlyMovements:
          type: integer
          format: int32
          example: 15

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2023-07-15T12:00:00Z"
        status:
          type: integer
          format: int32
          example: 404
        error:
          type: string
          example: "Not Found"
        message:
          type: string
          example: "Account not found with id: 60c73def2aa3a75f5c5d0f7c"